added caching

+++ src/controllers/item/delete-item.controller.ts
@@ -1,17 +1,32 @@
 import Item from '@/models/item.model';
-import { TErrorResponse, TItemModel } from '@/types/types';
+import Menu from '@/models/menu.model';
+import { TErrorResponse, TItemModel, TMenuModel } from '@/types/types';
 
 type TItemInput = {
     userId: string;
     projectId: string; 
+    menuId: string; 
     id: string;
 }
-export default async function DeleteItem(sectionInput: TItemInput): Promise<TErrorResponse | {}> {
+
+type TItemOutput = {
+    menu: {
+        id: string;
+        code: string;
+    }|null;
+}
+
+export default async function DeleteItem(itemInput: TItemInput): Promise<TErrorResponse | TItemOutput> {
     try {
-        const {userId, projectId, id} = sectionInput;
+        const {userId, projectId, menuId, id} = itemInput;
 
         if (!id) {
             throw new Error('invalid request');
+        }
+
+        const menu: TMenuModel|null = await Menu.findOne({userId, projectId, _id: menuId});
+        if (!menu) {
+            throw new Error('invalid menu');
         }
 
         const item: TItemModel|null  = await Item.findOneAndDelete({
@@ -23,7 +38,12 @@ export default async function DeleteItem(sectionInput: TItemInput): Promise<TErr
             throw new Error('invalid item');
         }
 
-        return {};
+        return {
+            menu: {
+                id: menu.id,
+                code: menu.code
+            }
+        };
     } catch (error) {
         throw error;
     }
